feat(connect): separate Connect and create ConnectCore without web3
allows developers to use uport-connect without the large web3 dependency, whether with dist file or

if developer creating their own build
