+++ package.json
@@ -92,6 +92,7 @@
   "dependencies": {
     "async": "^2.1.4",
     "ethjs-abi": "^0.1.8",
+    "ethjs-provider-http": "^0.1.6",
     "ethjs-util": "^0.1.3",
     "ipfs-mini": "^1.0.1",
     "json-loader": "^0.5.4",
