+++ src/index.js
@@ -1,5 +1,6 @@
 import Connect from './Connect'
+import ConnectCore from './ConnectCore'
 import { getQRDataURI, closeQr, openQr } from './util/qrdisplay'
 const QRUtil = { getQRDataURI, closeQr, openQr }
 import { SimpleSigner, Credentials } from 'uport'
-export { Connect, QRUtil, SimpleSigner, Credentials }
+export { Connect, ConnectCore, QRUtil, SimpleSigner, Credentials }
